backend/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.15,<4.0
channels>=4.0,<5.0
Flask>=3.0,<4.0
requests>=2.31,<3.0
fastmcp>=2.13,<3.0
numpy>=1.26,<3.0
opencv-python>=4.9,<5.0
Pillow>=10.0,<12.0
pyzbar>=0.1.9,<1.0
mcp>=1.0.0
openai>=1.55.0,<2.0
python-dotenv>=1.0.0,<2.0
pydantic>=2.12.0,<3.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
edge-tts>=7.0.0,<8.0.0
